Se crea archivo cli.js donde se importa la funcion mdLinks y se le dan los argumentos para su llamado close #19
